tools/requirements.txt
numpy
python-dotenv
requests
beautifulsoup4
